requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
cloudinary==1.41.0
coverage==7.6.1
crispy-bootstrap4==2024.1
Django==5.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-crispy-forms==2.3
django-db-connection-pool==1.2.5
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-filter==24.3
django-formtools==2.5.1
django-js-asset==2.2.0
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-taggit==6.0.0
django-tinymce==4.1.0
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.8
model-mommy==2.0.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.9
PyJWT==2.9.0
pypng==0.20220715.0
pytz==2024.1
qrcode==7.4.2
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.36
sqlparams==6.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
whitenoise==6.7.0
